views: transit-visualization, custom trip stop diff icon layout
Stop time icons happen to align with the above change icons.
New layout from UX to avoid impression that the change-icon's change
count would apply only for the day below it. When it actually displays
the difference of the two days.
The problem is not just as bad for the view's other tables, because
there the change-icon row has more than just two indicators.