update linkify component to produce an external link icon if the detected link target points to a website outside NAP
This relaxes the external link detection a bit since it did not work for absolute paths. It is also possible to override the icon generation and a rudimentary whitelisting of domains is provided as well. The whitelisting is based on URL prefixes which may not be performant in the long run but should be just fine for now.