Make the application authority group id aware
This is hardcoded to keep the control over assignment on the backend; frontend can still fake this, but just by changing the id won't actually grant any special permissions, as the backend expects the one from configuration. This does introduce a slight race condition, the group needs to be created before deployment.