leave a trace about updating Leiningen from 2.7.1 (!!) to 2.9.10
The build was broken because JDK on bastion has been updated but tooling wasn't, now everything's fine again. Of course these Ansible scripts are out-of-date anyway and the actual change was done with sudo+curl, but at least this way there's a slight change someone will see this was updated.